Address strict compiler warnings, support double processing

+++ CarveDSP/CarveParameters.h
@@ -46,7 +46,7 @@ public:
  *      maximum value,
  *      default value
  */
-const ParameterDefinition::RangedParameter<float>   PREGAIN(0, 2, 1),
+const ParameterDefinition::RangedParameter<double>   PREGAIN(0, 2, 1),
                                                     POSTGAIN(0, 2, 0.5),
                                                     TWEAK(-1, 1, 0);
 //@}
